Changes...
Bump GL 4.1 to 4.5
Adds Enable GL Debug to Preferences dialog
Fixes for texture initialization and maintenance
Better shader status checking
Adds Group to Preset option to group tab buttons + Hotkey F4

Fixes or addresses these issues...

Dynamic-precision Camera2D in Trudy #101
Spline path perspective #100
Add timelime editor to window menu #99
Keyframe parsing #98 Warns if keyframe is incomplete
Copying dependencies for multi include #96
ColorWidget and FloatColorWidget bug #93
OpenEXR Bug #85

+++ Fragmentarium-Source/Fragmentarium/Parser/Preprocessor.cpp
@@ -92,17 +92,17 @@ void setSliderType(GuiParameter *p, QString sliderTypeString)
 
 glm::dvec4 parseQVector4D(QString s1, QString s2, QString s3, QString s4)
 {
-    return { parseFloat(s1), parseFloat(s2), parseFloat(s3), parseFloat(s4) };
+    return { parseDouble(s1), parseDouble(s2), parseDouble(s3), parseDouble(s4) };
 }
 
 glm::dvec3 parseQVector3D(QString s1, QString s2, QString s3)
 {
-    return { parseFloat(s1), parseFloat(s2), parseFloat(s3) };
+    return { parseDouble(s1), parseDouble(s2), parseDouble(s3) };
 }
 
 glm::dvec2 parseQVector2D(QString s1, QString s2)
 {
-    return { parseFloat(s1), parseFloat(s2) };
+    return { parseDouble(s1), parseDouble(s2) };
 }
 }
